+++ CHANGELOG.md
@@ -96,6 +96,7 @@
     - Created textAreaInputForm.tsx to create an ff7 inspired input form that is also a textarea and handles all the intricaacies of newlines with shift+enter or submitting with enter.
     - Changed robot page to show a random robot inspired by Amazon random puppies.
     - Minor styling changes to projects route.
+    - Removed tutorial route from main menu and menu header.
 
 
 ## Bug Fixes: